csUTh

csUTh calculates closed-system Th-230/U ages, including detrital correction.

csUTh(
  input_data,
  sample_name = "YP003",
  nbitchoice = 100,
  detcorrectionchoice = TRUE,
  R28det = 0.8,
  R28det_err = 0.08,
  R08det = 1,
  R08det_err = 0.05,
  R48det = 1,
  R48det_err = 0.02,
  keepfiltereddata = FALSE,
  print_summary = TRUE,
  with_plots = TRUE,
  save_plots = TRUE,
  save_output = FALSE
)

Arguments

input_data
Input data frame. The following columns need to be present in this data frame, with these exact names: Sample_ID, U234_U238_CORR, U234_U238_CORR_Int2SE, Th230_U238_CORR, Th230_U238_CORR_Int2SE, Th232_U238_CORR, Th232_U238_CORR_Int2SE.

sample_name
Name of the sample to calculate closed-system ages for. The string entered must match characters for the chosen sample in the column 'Sample_ID' of the data file. Default: 'MK16'.

nbitchoice
Number of iterations in the model. Recommended to have at least 100. Default: 100.

detcorrectionchoice
Do a detrital correction? Enter TRUE for yes, or FALSE for no. Default: TRUE

R28det
(232Th/238U) activity ratio of the detritus. Default: 0.8

R28det_err
Error on the (232Th/238U) activity ratio of the detritus. Default: 0.08

R08det
(230Th/238U) activity ratio of the detritus. Default: 1

R08det_err
Error on the (230Th/238U) activity ratio of the detritus. Default: 0.05

R48det
(234U/238U) activity ratio of the detritus. Default: 1

R48det_err
Error on the (234U/238U) activity ratio of the detritus. Default: 0.02

keepfiltereddata
Save filtered data on which an outlier test was performed? Only recommended if all analyses of a same sample are supposed to give the same age. Enter TRUE for yes, or FALSE for no. Default: FALSE

print_summary
Print a summary of the output to the console? Default: TRUE

with_plots
Draw plots? Default: TRUE

save_plots
Save plots as a png file to the current working directory? Default: TRUE

save_output
Save output data as a CSV file to the current working directory? Default: FALSE

Examples

data("Pan2018")
# Solve for sample YP003
output <- csUTh(Pan2018,
sample_name = 'YP003',
nbitchoice = 100,
detcorrectionchoice = TRUE,
keepfiltereddata = FALSE,
print_summary = TRUE,
with_plots = FALSE,
save_plots = FALSE)

#> All required columns are present in the input data. 

#> [1] "Mean age:  117.3 +/- 3.8  ka"