osUTh

osUTh

osUTh(
  input_data,
  nbit = 1,
  fsum_target = 0.01,
  U48_0_min = 1.265,
  U48_0_max = 1.275,
  l = 5.35,
  U_0 = 25,
  K_min = 1e-13,
  K_max = 1e-11,
  T_min = 1000,
  T_max = 20000,
  print_summary = TRUE,
  with_plots = TRUE,
  save_plots = TRUE,
  save_output = FALSE
)

Arguments

input_data
A data frame with the required columns

nbit
The number of iterations

fsum_target
The sum of the squared differences between the calculated and observed activity ratios.

U48_0_min
The minimum value allowed for the (^234^U/^238^U) activity ratio at the surface of the sample

U48_0_max
The maximum value allowed for the (^234^U/^238^U) activity ratio at the surface of the sample

l
The thickness of the sample in centimeters

U_0
The uranium concentration at the surface in ppm

K_min
The minimum value allowed for the uranium diffusion coefficient (in cm^2^/s)

K_max
The maximum value allowed for the uranium diffusion coefficient (in cm^2^/s)

T_min
The minimum value for the age of the specimen (yr)

T_max
The maximum value for the age of the specimen (yr)

print_summary
Print a summary of the output to the console? Default is TRUE

with_plots
Display a panel of plots of the output? Default is TRUE

save_plots
Save plots as a png file to the current working directory? Default: TRUE

save_output
Save output data as a CSV file to the current working directory? Default: FALSE

Value

A list of results

Examples

data("Hobbit_MH2T_for_iDAD")
output <- osUTh(Hobbit_MH2T_for_iDAD,
nbit = 1,
fsum_target = 0.01,
U48_0_min = 1.265,
U48_0_max = 1.275,
l = 5.35,
U_0 = 25,
K_min = 1e-13,
K_max = 1e-11,
T_min = 1e3,
T_max = 20e3,
print_summary = TRUE,
with_plots = FALSE,
save_plots = FALSE)

#> All required columns are present in the input data. 

#> [1] "Age: 6.3 +0/-0 ka"